ref: move network metrics out of tests
